Missed one....
Finish up the switch over to 'atexit' rather than 'on_exit', as it is compatible with more systems. This also meant we need to store the exit code somwhere to make it possible for the exit handler to examine that code.  To make it easy, created a macro EXIT(x) that can be used to invoke exit that will automatically save the exit status in a global for use by the 'Goodbye' routine


git-svn-id: https://smartmontools.svn.sourceforge.net/svnroot/smartmontools/trunk@1091 4ea69e1a-61f1-4043-bf83-b5c94c648137